Manual merge of changes from 92-improve-application-startup which was branched before final Mjolnir release.
