+++ src/main/java/edu/jhuapl/trinity/App.java
@@ -3,90 +3,35 @@
 package edu.jhuapl.trinity;
 
 import edu.jhuapl.trinity.css.StyleResourceProvider;
-import edu.jhuapl.trinity.data.Dimension;
-import edu.jhuapl.trinity.data.FactorLabel;
-import edu.jhuapl.trinity.data.files.ClusterCollectionFile;
-import edu.jhuapl.trinity.data.files.FeatureCollectionFile;
-import edu.jhuapl.trinity.data.files.ManifoldDataFile;
-import edu.jhuapl.trinity.data.messages.xai.FeatureCollection;
-import edu.jhuapl.trinity.data.messages.xai.FeatureVector;
 import edu.jhuapl.trinity.javafx.components.MatrixOverlay;
-import edu.jhuapl.trinity.javafx.components.panes.AnalysisLogPane;
-import edu.jhuapl.trinity.javafx.components.panes.HyperdrivePane;
-import edu.jhuapl.trinity.javafx.components.panes.ImageInspectorPane;
-import edu.jhuapl.trinity.javafx.components.panes.NavigatorPane;
-import edu.jhuapl.trinity.javafx.components.panes.PixelSelectionPane;
-import edu.jhuapl.trinity.javafx.components.panes.Shape3DControlPane;
-import edu.jhuapl.trinity.javafx.components.panes.TextPane;
-import edu.jhuapl.trinity.javafx.components.panes.TrajectoryTrackerPane;
 import edu.jhuapl.trinity.javafx.components.panes.VideoPane;
-import edu.jhuapl.trinity.javafx.components.panes.WaveformPane;
 import edu.jhuapl.trinity.javafx.components.radial.CircleProgressIndicator;
 import edu.jhuapl.trinity.javafx.components.radial.MainNavMenu;
 import edu.jhuapl.trinity.javafx.components.radial.ProgressStatus;
-import edu.jhuapl.trinity.javafx.components.timeline.Item;
-import edu.jhuapl.trinity.javafx.components.timeline.MissionTimerX;
-import edu.jhuapl.trinity.javafx.components.timeline.MissionTimerXBuilder;
-import edu.jhuapl.trinity.javafx.components.timeline.TimelineAnimation;
 import edu.jhuapl.trinity.javafx.events.ApplicationEvent;
 import edu.jhuapl.trinity.javafx.events.CommandTerminalEvent;
-import edu.jhuapl.trinity.javafx.events.FeatureVectorEvent;
 import edu.jhuapl.trinity.javafx.events.FullscreenEvent;
-import edu.jhuapl.trinity.javafx.events.GaussianMixtureEvent;
-import edu.jhuapl.trinity.javafx.events.HitEvent;
-import edu.jhuapl.trinity.javafx.events.ManifoldEvent;
-import edu.jhuapl.trinity.javafx.events.RestEvent;
-import edu.jhuapl.trinity.javafx.events.SearchEvent;
-import edu.jhuapl.trinity.javafx.events.SemanticMapEvent;
-import edu.jhuapl.trinity.javafx.events.ShapleyEvent;
-import edu.jhuapl.trinity.javafx.events.TimelineEvent;
-import edu.jhuapl.trinity.javafx.events.TrajectoryEvent;
+import edu.jhuapl.trinity.javafx.events.MissionTimerXEvent;
 import edu.jhuapl.trinity.javafx.events.ZeroMQEvent;
-import edu.jhuapl.trinity.javafx.handlers.FeatureVectorEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.GaussianMixtureEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.HitEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.ManifoldEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.RestEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.SearchEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.SemanticMapEventHandler;
-import edu.jhuapl.trinity.javafx.handlers.ShapleyEventHandler;
-import edu.jhuapl.trinity.javafx.javafx3d.Hyperspace3DPane;
-import edu.jhuapl.trinity.javafx.javafx3d.Hypersurface3DPane;
-import edu.jhuapl.trinity.javafx.javafx3d.Manifold3D;
-import edu.jhuapl.trinity.javafx.javafx3d.Projections3DPane;
-import edu.jhuapl.trinity.javafx.javafx3d.ProjectorPane;
 import edu.jhuapl.trinity.javafx.javafx3d.RetroWavePane;
-import edu.jhuapl.trinity.messages.CommandTask;
-import edu.jhuapl.trinity.messages.MessageProcessor;
 import edu.jhuapl.trinity.messages.RestAccessLayer;
-import edu.jhuapl.trinity.messages.ZeroMQFeedManager;
-import edu.jhuapl.trinity.messages.ZeroMQSubscriberConfig;
-import edu.jhuapl.trinity.utils.AnalysisUtils;
 import edu.jhuapl.trinity.utils.Configuration;
-import edu.jhuapl.trinity.utils.PCAConfig;
-import edu.jhuapl.trinity.utils.ResourceUtils;
-import edu.jhuapl.trinity.utils.VisibilityMap;
-import edu.jhuapl.trinity.utils.umap.Umap;
+import edu.jhuapl.trinity.utils.Utils;
 import javafx.animation.Animation;
 import javafx.animation.FadeTransition;
 import javafx.animation.KeyFrame;
 import javafx.animation.KeyValue;
 import javafx.animation.Timeline;
 import javafx.application.Application;
-import javafx.application.Platform;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleIntegerProperty;
-import javafx.concurrent.Task;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
-import javafx.scene.input.DragEvent;
-import javafx.scene.input.Dragboard;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
-import javafx.scene.input.TransferMode;
 import javafx.scene.layout.Background;
 import javafx.scene.layout.BackgroundFill;
 import javafx.scene.layout.BorderPane;
@@ -102,11 +47,7 @@ import lit.litfx.controls.output.AnimatedText;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 public class App extends Application {
@@ -125,56 +66,27 @@ public class App extends Application {
     Timeline intro;
     MatrixOverlay matrixOverlay;
     RetroWavePane retroWavePane = null;
-    Hyperspace3DPane hyperspace3DPane;
-    Hypersurface3DPane hypersurface3DPane;
-    ProjectorPane projectorPane;
-    Projections3DPane projections3DPane;
-    TrajectoryTrackerPane trajectoryTrackerPane;
-    TextPane textConsolePane;
-    VideoPane videoPane;
-    NavigatorPane navigatorPane;
-    WaveformPane waveformPane;
-    Shape3DControlPane shape3DControlPane;
-    AnalysisLogPane analysisLogPane;
-    PixelSelectionPane pixelSelectionPane;
-    ImageInspectorPane imageInspectorPane;
-    HyperdrivePane hyperdrivePane;
     CircleProgressIndicator circleSpinner;
     //Command line argument support
     Map<String, String> namedParameters;
-    List<String> unnamedParameters;
     AnimatedText animatedConsoleText;
-    MessageProcessor processor;
-    ZeroMQSubscriberConfig subscriberConfig;
-    ZeroMQFeedManager feed;
-    ManifoldEventHandler meh;
-    FeatureVectorEventHandler fveh;
-    ShapleyEventHandler sheh;
-    GaussianMixtureEventHandler gmeh;
-    SemanticMapEventHandler smeh;
-    SearchEventHandler seh;
-    HitEventHandler heh;
-    RestEventHandler reh;
 
-    boolean hyperspaceIntroShown = false;
-    boolean hypersurfaceIntroShown = false;
     boolean matrixShowing = false;
     boolean enableMatrix = false;
     static boolean matrixEnabled = false;
-    MissionTimerX missionTimerX;
-    TimelineAnimation timelineAnimation;
     boolean is4k = false;
     boolean enableHttp = false;
 
     @Override
     public void start(Stage stage) throws IOException {
+        System.out.println("App.start() begin... ");
+        long startTime = System.nanoTime();
         LOG.info("Attempting to read defaults...");
         try {
             LOG.info("Build Date: {}", Configuration.getBuildDate());
         } catch (Exception ex) {
             LOG.error(null, ex);
         }
-        //theConfig = Configuration.defaultConfiguration();
         LOG.info("Starting JavaFX rendering...");
         centerStack = new StackPane();
         centerStack.setBackground(transBack);
@@ -183,7 +95,6 @@ public class App extends Application {
         animatedConsoleText.setStyle("    -fx-font-size: 30;");
         StackPane.setAlignment(animatedConsoleText, Pos.BOTTOM_LEFT);
         animatedConsoleText.setTranslateY(-100);
-        centerStack.getChildren().add(animatedConsoleText);
 
         LOG.info("Building Scene Stack...");
         BorderPane bp = new BorderPane(centerStack);
@@ -191,7 +102,6 @@ public class App extends Application {
         bp.getStyleClass().add("trinity-pane");
         Scene scene = new Scene(bp, Color.BLACK);
         stage.setScene(scene);
-        stage.show();
 
         LOG.info("Styling Scene and Stage...");
         //animatedConsoleText.animate("Styling Scene and Stage...");
@@ -215,215 +125,14 @@ public class App extends Application {
         //@HACK This ultra dirty hack lets me easily add floating panes from anywhere
         pathPane = desktopPane;
         theScene = scene;
-        videoPane = new VideoPane(scene, desktopPane);
-        theVideo = videoPane;
-        matrixEnabled = enableMatrix;
         //</HACK>
+
         LOG.info("Building menu system...");
-        //animatedConsoleText.animate("Building menu system...");
         mainNavMenu = new MainNavMenu(scene);
         StackPane.setAlignment(mainNavMenu, Pos.BOTTOM_RIGHT);
         mainNavMenu.hideRadialMenu();
         mainNavMenu.setTranslateX(-mainNavMenu.getInnerRadius());
         mainNavMenu.setTranslateY(-mainNavMenu.getInnerRadius());
-
-        LOG.info("Constructing 3D subscenes...");
-        //animatedConsoleText.animate("Constructing 3D subscenes...");
-        projections3DPane = new Projections3DPane(scene);
-        projections3DPane.setVisible(false); //start off hidden
-        hypersurface3DPane = new Hypersurface3DPane(scene);
-        hypersurface3DPane.setVisible(false); //start off hidden
-        hyperspace3DPane = new Hyperspace3DPane(scene);
-        hyperspace3DPane.setVisible(false); //start off hidden
-        navigatorPane = new NavigatorPane(scene, pathPane);
-
-        projectorPane = new ProjectorPane();
-        projectorPane.setVisible(false); //start off hidden
-        analysisLogPane = new AnalysisLogPane(scene, desktopPane);
-        pixelSelectionPane = new PixelSelectionPane(scene, pathPane);
-        imageInspectorPane = new ImageInspectorPane(scene, pathPane);
-        hyperdrivePane = new HyperdrivePane(scene, pathPane);
-
-        LOG.info("Registering Event Handlers...");
-        //animatedConsoleText.animate("Registering Event Handlers...");
-        scene.addEventHandler(FeatureVectorEvent.REQUEST_FEATURE_COLLECTION, event -> {
-            FeatureCollection fc = new FeatureCollection();
-            fc.setFeatures(hyperspace3DPane.getAllFeatureVectors());
-            hypersurface3DPane.addFeatureCollection(fc, false);
-        });
-
-        scene.addEventHandler(ManifoldEvent.GENERATE_NEW_UMAP, event -> {
-            Platform.runLater(() -> {
-                ProgressStatus ps = new ProgressStatus("Generating UMAP Settings...", -1);
-                scene.getRoot().fireEvent(
-                    new ApplicationEvent(ApplicationEvent.SHOW_BUSY_INDICATOR, ps));
-            });
-            Umap umap = AnalysisUtils.getDefaultUmap();
-            if (null != event.object1)
-                umap = (Umap) event.object1;
-            else if (null != projections3DPane.latestUmap)
-                umap = projections3DPane.latestUmap;
-            umap.setThreads(24);
-            ManifoldEvent.POINT_SOURCE source = ManifoldEvent.POINT_SOURCE.HYPERSPACE;
-            if (null != event.object2)
-                source = (ManifoldEvent.POINT_SOURCE) event.object2;
-            FeatureCollection originalFC = getFeaturesBySource(source);
-            projections3DPane.projectFeatureCollection(originalFC, umap);
-        });
-
-        scene.addEventHandler(ManifoldEvent.GENERATE_NEW_PCA, event -> {
-            Platform.runLater(() -> {
-                ProgressStatus ps = new ProgressStatus("Projecting PCA...", -1);
-                scene.getRoot().fireEvent(
-                    new ApplicationEvent(ApplicationEvent.SHOW_BUSY_INDICATOR, ps));
-            });
-            PCAConfig config = (PCAConfig) event.object1;
-            ManifoldEvent.POINT_SOURCE source = (ManifoldEvent.POINT_SOURCE) event.object2;
-            FeatureCollection originalFC = getFeaturesBySource(source);
-            projections3DPane.projectFeatureCollection(originalFC, config);
-        });
-
-        scene.addEventHandler(FeatureVectorEvent.PROJECT_SURFACE_GRID, event -> {
-            FeatureCollection originalFC = new FeatureCollection();
-            originalFC.setFeatures(hypersurface3DPane.getAllFeatureVectors());
-            double[][] data = AnalysisUtils.fitUMAP(originalFC);
-            FeatureCollection projectedFC = FeatureCollection.fromData(data);
-            for (int i = 0; i < originalFC.getFeatures().size(); i++) {
-                FeatureVector origFV = originalFC.getFeatures().get(i);
-                projectedFC.getFeatures().get(i).setLabel(origFV.getLabel());
-                projectedFC.getFeatures().get(i).setScore(origFV.getScore());
-                projectedFC.getFeatures().get(i).setImageURL(origFV.getImageURL());
-            }
-            if (null != originalFC.getDimensionLabels()) {
-                projectedFC.setDimensionLabels(originalFC.getDimensionLabels());
-                projections3DPane.setDimensionLabels(originalFC.getDimensionLabels());
-            }
-            projections3DPane.setHyperDimensionFeatures(originalFC);
-            projections3DPane.addFeatureCollection(projectedFC, false);
-
-        });
-
-        scene.addEventHandler(FeatureVectorEvent.PROJECT_FEATURE_COLLECTION, event -> {
-            FeatureCollection originalFC = new FeatureCollection();
-            originalFC.setFeatures(hyperspace3DPane.getAllFeatureVectors());
-            double[][] data = AnalysisUtils.fitUMAP(originalFC);
-            FeatureCollection projectedFC = FeatureCollection.fromData(data);
-            for (int i = 0; i < originalFC.getFeatures().size(); i++) {
-                FeatureVector origFV = originalFC.getFeatures().get(i);
-                projectedFC.getFeatures().get(i).setLabel(origFV.getLabel());
-                projectedFC.getFeatures().get(i).setScore(origFV.getScore());
-                projectedFC.getFeatures().get(i).setImageURL(origFV.getImageURL());
-            }
-            if (null != originalFC.getDimensionLabels()) {
-                projectedFC.setDimensionLabels(originalFC.getDimensionLabels());
-                projections3DPane.setDimensionLabels(originalFC.getDimensionLabels());
-            }
-            projections3DPane.setHyperDimensionFeatures(originalFC);
-            projections3DPane.addFeatureCollection(projectedFC, false);
-        });
-
-        scene.addEventHandler(FeatureVectorEvent.EXPORT_FEATURE_COLLECTION, event -> {
-            File file = (File) event.object;
-            FeatureCollection fc = new FeatureCollection();
-            fc.setFeatures(hyperspace3DPane.getAllFeatureVectors());
-            try {
-                FeatureCollectionFile fcf = new FeatureCollectionFile(file.getAbsolutePath(), false);
-                fcf.featureCollection = fc;
-                fcf.writeContent();
-            } catch (IOException ex) {
-                LOG.error(null, ex);
-            }
-            event.consume();
-        });
-        //some helpful handling for the main view
-        centerStack.addEventHandler(DragEvent.DRAG_OVER, event -> {
-            if (ResourceUtils.canDragOver(event)) {
-                event.acceptTransferModes(TransferMode.COPY);
-            } else {
-                event.consume();
-            }
-        });
-        centerStack.addEventHandler(DragEvent.DRAG_DROPPED, e -> {
-            String dropType = ResourceUtils.detectDropType(e);
-            if (!dropType.contentEquals("UNKNOWN")) {
-                if (ResourceUtils.promptUserOnCommand(dropType)) {
-                    try {
-                        switch (dropType) {
-                            case "Hyperspace" -> CommandTask.execute(scene, "VIEW_HYPERSPACE", 0, null);
-                            case "Hypersurface" -> CommandTask.execute(scene, "VIEW_HYPERSURFACE", 0, null);
-                            case "Projections" -> CommandTask.execute(scene, "VIEW_PROJECTIONS", 0, null);
-                        }
-                    } catch (InterruptedException ex) {
-                        LOG.error(null, ex);
-                    }
-                }
-            }
-            ResourceUtils.onDragDropped(e, scene);
-        });
-        ////////////////////////
-
-        hyperspace3DPane.addEventHandler(DragEvent.DRAG_OVER, event -> {
-            if (ResourceUtils.canDragOver(event)) {
-                event.acceptTransferModes(TransferMode.COPY);
-            } else {
-                event.consume();
-            }
-        });
-        hyperspace3DPane.addEventHandler(DragEvent.DRAG_DROPPED,
-            e -> ResourceUtils.onDragDropped(e, scene));
-        hypersurface3DPane.addEventHandler(DragEvent.DRAG_OVER, event -> {
-            if (ResourceUtils.canDragOver(event)) {
-                event.acceptTransferModes(TransferMode.COPY);
-            } else {
-                event.consume();
-            }
-        });
-        hypersurface3DPane.addEventHandler(DragEvent.DRAG_DROPPED,
-            e -> ResourceUtils.onDragDropped(e, scene));
-
-        projections3DPane.addEventHandler(DragEvent.DRAG_OVER, event -> {
-            if (ResourceUtils.canDragOver(event)) { // && projections3DPane.autoProjectionProperty.get()) {
-                event.acceptTransferModes(TransferMode.COPY);
-            } else {
-                event.consume();
-            }
-        });
-        projections3DPane.addEventHandler(DragEvent.DRAG_DROPPED, e -> {
-            Dragboard db = e.getDragboard();
-            if (db.hasFiles()) {
-                for (File file : db.getFiles()) {
-                    try {
-                        if (projections3DPane.autoProjectionProperty.get()) {
-                            if (FeatureCollectionFile.isFeatureCollectionFile(file)) {
-                                FeatureCollectionFile fcFile = new FeatureCollectionFile(
-                                    file.getAbsolutePath(), true);
-                                fveh.scanLabelsAndLayers(fcFile.featureCollection.getFeatures());
-                                projections3DPane.transformFeatureCollection(fcFile.featureCollection);
-                            }
-                        }
-                        if (ClusterCollectionFile.isClusterCollectionFile(file)) {
-                            ClusterCollectionFile ccFile = new ClusterCollectionFile(file.getAbsolutePath(), true);
-                            scene.getRoot().fireEvent(
-                                new ManifoldEvent(ManifoldEvent.NEW_CLUSTER_COLLECTION, ccFile.clusterCollection));
-                        }
-                        if (ManifoldDataFile.isManifoldDataFile(file)) {
-                            ManifoldDataFile mdFile = new ManifoldDataFile(file.getAbsolutePath(), true);
-                            Platform.runLater(() -> scene.getRoot().fireEvent(
-                                new ManifoldEvent(ManifoldEvent.NEW_MANIFOLD_DATA, mdFile.manifoldData)));
-                        }
-                    } catch (IOException ex) {
-                        LOG.error(null, ex);
-                    }
-                }
-            }
-        });
-
-        //Add the base main tools
-        //insert before animated console text
-        centerStack.getChildren().add(0, hyperspace3DPane);
-        centerStack.getChildren().add(0, projections3DPane);
-        centerStack.getChildren().add(0, hypersurface3DPane);
-        centerStack.getChildren().add(0, projectorPane);
 
         LOG.info("Parsing command line...");
         //animatedConsoleText.animate("Parsing command line...");
@@ -498,36 +207,26 @@ public class App extends Application {
                 LOG.info("HTTP found: {}", httpValue);
                 enableHttp = Boolean.parseBoolean(httpValue);
             }
-
-        }
-        LOG.info("Setting up RestEventHandler...");
-        reh = new RestEventHandler(scene);
-        scene.addEventHandler(RestEvent.START_RESTSERVER_THREAD, reh);
-        scene.addEventHandler(RestEvent.TERMINATE_RESTSERVER_THREAD, reh);
-        if (enableHttp) {
-            reh.startHttpService();
         }
 
-        LOG.info("Setting up Mission Timer...");
-        setupMissionTimer(scene);
         //add helper tools and overlays
         circleSpinner = new CircleProgressIndicator();
         circleSpinner.setLabelLater("...Working...");
         circleSpinner.defaultOpacity = 1.0;
         circleSpinner.setOpacity(0.0); ///instead of setVisible(false)
 
-        centerStack.getChildren().remove(animatedConsoleText);
-        centerStack.getChildren().addAll(desktopPane, mainNavMenu,
-            missionTimerX, circleSpinner);
+        centerStack.getChildren().addAll(desktopPane, 
+            //mainNavMenu, 
+            circleSpinner);
         centerStack.getChildren().add(animatedConsoleText);
-
+        stage.show();        
 
         LOG.info("Setting up Matrix Digital Rain...");
         // fun matrix effect, use Alt + N
         matrixOverlay = new MatrixOverlay(scene, centerStack);
         bp.setOnSwipeUp(e -> {
             if (enableMatrix) {
-                if (e.getTouchCount() > 2) {
+                if (e.getTouchCount() > 2) { //three finger swipe
                     MatrixOverlay.on.set(false);
                     matrixOverlay.matrixOff.run();
                     e.consume(); // <-- stops passing the event to next node
@@ -536,14 +235,13 @@ public class App extends Application {
         });
         bp.setOnSwipeDown(e -> {
             if (enableMatrix) {
-                if (e.getTouchCount() > 2) {
+                if (e.getTouchCount() > 2) { //three finger swipe
                     MatrixOverlay.on.set(true);
                     matrixOverlay.matrixOn.run();
                     e.consume(); // <-- stops passing the event to next node
                 }
             }
         });
-
         LOG.info("Adding Event handlers... ");
         scene.addEventHandler(ApplicationEvent.SHOW_BUSY_INDICATOR, e -> {
             if (null == e.object) {
@@ -565,318 +263,6 @@ public class App extends Application {
         });
 
         scene.addEventHandler(ApplicationEvent.SHUTDOWN, e -> shutdown(false));
-        LOG.info("Text Console ");
-        scene.addEventHandler(ApplicationEvent.SHOW_TEXT_CONSOLE, e -> {
-            if (null == textConsolePane) {
-                textConsolePane = new TextPane(scene, pathPane);
-                textConsolePane.setOpaqueEnabled(true);
-            }
-            if (!pathPane.getChildren().contains(textConsolePane)) {
-                pathPane.getChildren().add(textConsolePane);
-                textConsolePane.slideInPane();
-            } else {
-                textConsolePane.show();
-            }
-            if (null != e.object) {
-                Platform.runLater(() -> textConsolePane.setText((String) e.object));
-            }
-        });
-        LOG.info("Video Pane ");
-        scene.addEventHandler(ApplicationEvent.SHOW_VIDEO_PANE, e -> {
-            if (null == videoPane) {
-                videoPane = new VideoPane(scene, pathPane);
-                theVideo = videoPane;
-            }
-            if (!pathPane.getChildren().contains(videoPane)) {
-                pathPane.getChildren().add(videoPane);
-                videoPane.setTranslateX(desktopPane.getWidth() / 2.0
-                    - videoPane.getBoundsInLocal().getWidth() / 2.0);
-                videoPane.setTranslateY(desktopPane.getHeight() / 2.0
-                    - videoPane.getBoundsInLocal().getHeight() / 2.0);
-                videoPane.slideInPane();
-            } else {
-                videoPane.show();
-            }
-            if (null != e.object) {
-                String title = (String) e.object;
-                videoPane.mainTitleTextProperty.set(title);
-            }
-            if (null != e.object2) {
-                String caption = (String) e.object2;
-                videoPane.mainTitleText2Property.set(caption);
-            }
-            videoPane.setVideo();
-        });
-        LOG.info("Image Navigator ");
-        scene.addEventHandler(ApplicationEvent.SHOW_NAVIGATOR_PANE, e -> {
-            if (null == navigatorPane) {
-                navigatorPane = new NavigatorPane(scene, pathPane);
-            }
-            if (!pathPane.getChildren().contains(navigatorPane)) {
-                pathPane.getChildren().add(navigatorPane);
-                navigatorPane.slideInPane();
-            } else {
-                navigatorPane.show();
-            }
-            if (null != e.object) {
-                Platform.runLater(() -> navigatorPane.setImage((Image) e.object));
-            }
-        });
-        LOG.info("Analysis Log View");
-        scene.addEventHandler(ApplicationEvent.SHOW_ANALYSISLOG_PANE, e -> {
-            if (null == analysisLogPane) {
-                analysisLogPane = new AnalysisLogPane(scene, pathPane);
-            }
-            if (!pathPane.getChildren().contains(analysisLogPane)) {
-                pathPane.getChildren().add(analysisLogPane);
-                analysisLogPane.slideInPane();
-            } else {
-                analysisLogPane.show();
-            }
-        });
-        LOG.info("Pixel Selection Pane");
-        scene.addEventHandler(ApplicationEvent.SHOW_PIXEL_SELECTION, e -> {
-            if (null == pixelSelectionPane) {
-                pixelSelectionPane = new PixelSelectionPane(scene, pathPane);
-            }
-            if (!pathPane.getChildren().contains(pixelSelectionPane)) {
-                pathPane.getChildren().add(pixelSelectionPane);
-                pixelSelectionPane.slideInPane();
-            } else {
-                pixelSelectionPane.show();
-            }
-            if (null != e.object) {
-                Image image = (Image) e.object;
-                pixelSelectionPane.setImage(image);
-            }
-        });
-        LOG.info("Image Inspection Pane");
-        scene.addEventHandler(ApplicationEvent.SHOW_IMAGE_INSPECTION, e -> {
-            if (null != e.object) {
-                Boolean show = (Boolean) e.object;
-                if (show) {
-                    if (null == imageInspectorPane) {
-                        imageInspectorPane = new ImageInspectorPane(scene, pathPane);
-                    }
-                    if (!pathPane.getChildren().contains(imageInspectorPane)) {
-                        pathPane.getChildren().add(imageInspectorPane);
-                        imageInspectorPane.slideInPane();
-                    } else {
-                        imageInspectorPane.show();
-                    }
-                }
-            }
-        });
-        LOG.info("Waveform View ");
-        scene.addEventHandler(ApplicationEvent.SHOW_WAVEFORM_PANE, e -> {
-            if (null == waveformPane) {
-                waveformPane = new WaveformPane(scene, pathPane);
-            }
-            if (!pathPane.getChildren().contains(waveformPane)) {
-                pathPane.getChildren().add(waveformPane);
-                waveformPane.slideInPane();
-            } else {
-                waveformPane.show();
-            }
-            if (null != e.object) {
-                Platform.runLater(() -> waveformPane.setWaveform((File) e.object));
-            }
-        });
-        LOG.info("Trajectory Tracker ");
-        scene.addEventHandler(TrajectoryEvent.SHOW_TRAJECTORY_TRACKER, e -> {
-            if (null == trajectoryTrackerPane) {
-                trajectoryTrackerPane = new TrajectoryTrackerPane(scene, pathPane);
-            }
-            if (!pathPane.getChildren().contains(trajectoryTrackerPane)) {
-                pathPane.getChildren().add(trajectoryTrackerPane);
-                trajectoryTrackerPane.slideInPane();
-            } else {
-                trajectoryTrackerPane.show();
-            }
-        });
-        LOG.info("Projections ");
-        scene.addEventHandler(ApplicationEvent.SHOW_PROJECTIONS, e -> {
-            if (projections3DPane.isVisible()) {
-                projections3DPane.setVisible(false);
-                if (null != retroWavePane) {
-                    retroWavePane.animateShow();
-                }
-            } else {
-                if (null != retroWavePane) {
-                    retroWavePane.animateHide();
-                }
-                hyperspace3DPane.setVisible(false);
-                hypersurface3DPane.setVisible(false);
-                Platform.runLater(() -> {
-                    projections3DPane.setVisible(true);
-                });
-            }
-        });
-
-        scene.addEventHandler(ApplicationEvent.AUTO_PROJECTION_MODE, e -> {
-            boolean enabled = (boolean) e.object;
-            if (enabled) {
-                if (null != retroWavePane) {
-                    retroWavePane.animateHide();
-                }
-                hyperspace3DPane.setVisible(false);
-                hypersurface3DPane.setVisible(false);
-                projections3DPane.setVisible(true);
-            }
-            projections3DPane.enableAutoProjection(enabled);
-            hyperspace3DPane.enableAutoProjection(enabled);
-        });
-        LOG.info("Hypersurface ");
-        scene.addEventHandler(ApplicationEvent.SHOW_HYPERSURFACE, e -> {
-            if (hypersurface3DPane.isVisible()) {
-                Platform.runLater(() -> hypersurface3DPane.setVisible(false));
-                if (null != retroWavePane) {
-                    retroWavePane.animateShow();
-                }
-            } else {
-                if (null != retroWavePane) {
-                    retroWavePane.animateHide();
-                }
-                hyperspace3DPane.setVisible(false);
-                if (null != projections3DPane) {
-                    projections3DPane.setVisible(false);
-                }
-                fadeOutConsole(500);
-                if (!hypersurfaceIntroShown) {
-                    hypersurface3DPane.showFA3D();
-                    hypersurfaceIntroShown = true;
-                } else {
-                    Platform.runLater(() -> hypersurface3DPane.setVisible(true));
-                }
-            }
-        });
-        LOG.info("Hyperspace ");
-        scene.addEventHandler(ApplicationEvent.SHOW_HYPERSPACE, e -> {
-            if (hyperspace3DPane.isVisible()) {
-                hyperspace3DPane.setVisible(false);
-                if (null != retroWavePane) {
-                    retroWavePane.animateShow();
-                }
-                fadeOutConsole(500);
-            } else {
-                hypersurface3DPane.setVisible(false);
-                if (null != projections3DPane) {
-                    projections3DPane.setVisible(false);
-                }
-                if (null != retroWavePane) {
-                    retroWavePane.animateHide();
-                }
-                hyperspace3DPane.setVisible(true);
-                if (!hyperspaceIntroShown) {
-                    hyperspace3DPane.intro(1000);
-                    hyperspaceIntroShown = true;
-                }
-                fadeInConsole(500);
-            }
-        });
-        scene.addEventHandler(ApplicationEvent.SHOW_SHAPE3D_CONTROLS, e -> {
-            Manifold3D manifold3D = (Manifold3D) e.object;
-            if (null == shape3DControlPane) {
-                shape3DControlPane = new Shape3DControlPane(scene, pathPane);
-            }
-            if (null != manifold3D) {
-                shape3DControlPane.setShape3D(manifold3D);
-            }
-            if (!pathPane.getChildren().contains(shape3DControlPane)) {
-                pathPane.getChildren().add(shape3DControlPane);
-                shape3DControlPane.slideInPane();
-            } else {
-                shape3DControlPane.show();
-            }
-        });
-        scene.addEventHandler(ApplicationEvent.SHOW_PROJECTOR_PANE, e -> {
-            boolean visible = !projectorPane.isVisible();
-            if (null != e.object)
-                visible = (boolean) e.object;
-            projectorPane.setVisible(visible);
-            if (visible && projectorPane.firstTime == true) {
-                projectorPane.firstTime = false;
-                projectorPane.scanAndAnimate();
-            }
-        });
-        scene.addEventHandler(ApplicationEvent.SHOW_HYPERDRIVE_PANE, e -> {
-            if (!pathPane.getChildren().contains(hyperdrivePane)) {
-                pathPane.getChildren().add(hyperdrivePane);
-                hyperdrivePane.slideInPane();
-            } else {
-                hyperdrivePane.show();
-            }
-        });
-        LOG.info("Data Handlers ");
-        gmeh = new GaussianMixtureEventHandler();
-        scene.getRoot().addEventHandler(GaussianMixtureEvent.NEW_GAUSSIAN_MIXTURE, gmeh);
-        scene.getRoot().addEventHandler(GaussianMixtureEvent.LOCATE_GAUSSIAN_MIXTURE, gmeh);
-        scene.getRoot().addEventHandler(GaussianMixtureEvent.NEW_GAUSSIAN_COLLECTION, gmeh);
-        gmeh.addGaussianMixtureRenderer(hyperspace3DPane);
-
-        fveh = new FeatureVectorEventHandler();
-        scene.getRoot().addEventHandler(FeatureVectorEvent.NEW_FEATURE_VECTOR, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.LOCATE_FEATURE_VECTOR, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.NEW_FEATURE_COLLECTION, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.RESCAN_FACTOR_LABELS, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.RESCAN_FEATURE_LAYERS, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.NEW_LABEL_CONFIG, fveh);
-        scene.getRoot().addEventHandler(FeatureVectorEvent.CLEAR_ALL_FEATUREVECTORS, fveh);
-        fveh.addFeatureVectorRenderer(hyperspace3DPane);
-
-        meh = new ManifoldEventHandler();
-        scene.getRoot().addEventHandler(ManifoldEvent.NEW_UMAP_CONFIG, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.NEW_MANIFOLD_CLUSTER, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.NEW_MANIFOLD_DATA, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.EXPORT_MANIFOLD_DATA, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.CLEAR_ALL_MANIFOLDS, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.GENERATE_PROJECTION_MANIFOLD, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SET_SCALE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_ROTATE_X, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_ROTATE_Y, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_ROTATE_Z, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SET_YAWPITCHROLL, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_FILL_DRAWMODE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_LINE_DRAWMODE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_FRONT_CULLFACE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_BACK_CULLFACE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_NONE_CULLFACE, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SHOW_CONTROL, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SHOW_WIREFRAME, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SHOW_WIREFRAME, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_DIFFUSE_COLOR, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_SPECULAR_COLOR, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.MANIFOLD_WIREFRAME_COLOR, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.FIND_PROJECTION_CLUSTERS, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.NEW_CLUSTER_COLLECTION, meh);
-        scene.getRoot().addEventHandler(ManifoldEvent.NEW_PROJECTION_VECTOR, meh);
-        meh.addManifoldRenderer(projections3DPane);
-        scene.getRoot().addEventHandler(ManifoldEvent.FIND_HYPERSURFACE_CLUSTERS, e -> {
-            hypersurface3DPane.findClusters((ManifoldEvent.ProjectionConfig) e.object1);
-        });
-
-        sheh = new ShapleyEventHandler();
-        scene.getRoot().addEventHandler(ShapleyEvent.NEW_SHAPLEY_VECTOR, sheh);
-        scene.getRoot().addEventHandler(ShapleyEvent.NEW_SHAPLEY_COLLECTION, sheh);
-        sheh.addShapleyVectorRenderer(hypersurface3DPane);
-
-        heh = new HitEventHandler(desktopPane);
-        scene.getRoot().addEventHandler(HitEvent.PROJECTILE_HIT_SHAPE, heh);
-        scene.getRoot().addEventHandler(HitEvent.TRACKING_PROJECTILE_EVENTS, heh);
-
-        smeh = new SemanticMapEventHandler(false);
-        scene.getRoot().addEventHandler(SemanticMapEvent.NEW_SEMANTIC_MAP, smeh);
-        scene.getRoot().addEventHandler(SemanticMapEvent.NEW_SEMANTICMAP_COLLECTION, smeh);
-        smeh.addFeatureVectorRenderer(hyperspace3DPane);
-        smeh.addSemanticMapRenderer(hypersurface3DPane);
-
-        seh = new SearchEventHandler();
-        scene.getRoot().addEventHandler(SearchEvent.FILTER_BY_TERM, seh);
-        scene.getRoot().addEventHandler(SearchEvent.FILTER_BY_SCORE, seh);
-        scene.getRoot().addEventHandler(SearchEvent.CLEAR_ALL_FILTERS, seh);
-        scene.getRoot().addEventHandler(SearchEvent.FIND_BY_QUERY, seh);
-        scene.getRoot().addEventHandler(SearchEvent.QUERY_EMBEDDINGS_RESPONSE, seh);
-        seh.addFeatureVectorRenderer(hyperspace3DPane);
 
         scene.getRoot().addEventHandler(CovalentPaneEvent.COVALENT_PANE_CLOSE, covalentEvent -> {
             desktopPane.getChildren().remove(covalentEvent.pathPane);
@@ -923,139 +309,36 @@ public class App extends Application {
             animatedConsoleText.animate("> " + e.text);
         });
         scene.addEventHandler(CommandTerminalEvent.FADE_OUT, e -> fadeOutConsole(e.timeMS));
-        LOG.info("Establishing Messaging Feed...");
-        processor = new MessageProcessor(scene);
-        subscriberConfig = new ZeroMQSubscriberConfig(
-            "ZeroMQ Subscriber", "ZeroMQFeedManager.",
-            "tcp://localhost:5563", "ALL", "SomeIDValue", 250);
-        feed = new ZeroMQFeedManager(2, subscriberConfig, processor);
-        scene.getRoot().addEventHandler(ZeroMQEvent.ZEROMQ_ESTABLISH_CONNECTION, e -> {
-            subscriberConfig = e.subscriberConfig;
-            if (null != subscriberConfig) {
-                Task task = new Task() {
-                    @Override
-                    protected Void call() throws Exception {
-                        feed.setConfig(subscriberConfig);
-                        feed.startProcessing();
-                        return null;
-                    }
-                };
-                Thread thread = new Thread(task);
-                thread.setDaemon(true);
-                thread.start();
-            }
-        });
-        scene.getRoot().addEventHandler(ZeroMQEvent.ZEROMQ_TERMINATE_CONNECTION, e -> {
-            if (null != feed) {
-                feed.disconnect(false);
-            }
-        });
-        scene.getRoot().addEventHandler(ZeroMQEvent.ZEROMQ_START_PROCESSING, e -> {
-            feed.setEnableProcessing(true);
-        });
-        scene.getRoot().addEventHandler(ZeroMQEvent.ZEROMQ_STOP_PROCESSING, e -> {
-            feed.setEnableProcessing(false);
-        });
-        LOG.info("Checking to auto-enable ZeroMQ Listener...");
-        if (null != namedParameters && namedParameters.containsKey("zeromq")) {
-            String zeroMQ = namedParameters.get("zeromq");
-            if (null != zeroMQ) {
-                LOG.info("ZeroMQ Status: {}", zeroMQ);
-            }
-            Platform.runLater(() -> scene.getRoot().fireEvent(
-                new ZeroMQEvent(ZeroMQEvent.ZEROMQ_ESTABLISH_CONNECTION, subscriberConfig)));
-        }
 
         LOG.info("User Interface Lit...");
-        //animatedConsoleText.animate("User Inteface Lit...");
-        intro();
-        missionTimerX.updateTime(0, "1");
+        AppAsyncManager async = new AppAsyncManager(scene, centerStack, desktopPane, circleSpinner, namedParameters);
+        async.setOnSucceeded(s -> intro());
+        async.setOnFailed(s -> intro());            
+        Utils.printTotalTime(startTime);
+        Thread.startVirtualThread(async);
     }
-
-    private FeatureCollection getFeaturesBySource(ManifoldEvent.POINT_SOURCE source) {
-        FeatureCollection originalFC = new FeatureCollection();
-        if (source == ManifoldEvent.POINT_SOURCE.HYPERSURFACE) {
-            originalFC = FeatureCollection.fromData(hypersurface3DPane.dataGrid);
-            for (int i = 0; i < originalFC.getFeatures().size() - 1; i++) {
-                FeatureVector actualFV = hypersurface3DPane.getAllFeatureVectors().get(i);
-                originalFC.getFeatures().get(i).setLabel(actualFV.getLabel());
-                originalFC.getFeatures().get(i).setScore(actualFV.getScore());
-                originalFC.getFeatures().get(i).setImageURL(actualFV.getImageURL());
-            }
-        } else {
-            //filter to only use visible points
-//Original method only supports visibilty via labels
-//            originalFC.setFeatures(
-//                hyperspace3DPane.getAllFeatureVectors().stream().filter(fv -> {
-//                    return FactorLabel.visibilityByLabel(fv.getLabel());
-//                }).toList());
-
-            int size = hyperspace3DPane.getAllFeatureVectors().size();
-            List<FeatureVector> visibleFeatures = new ArrayList<>();
-            for (int i = 0; i < size; i++) {
-                if (FactorLabel.visibilityByLabel(
-                    hyperspace3DPane.getAllFeatureVectors().get(i).getLabel())
-                    && VisibilityMap.visibilityByIndex(i)) {
-                    visibleFeatures.add(hyperspace3DPane.getAllFeatureVectors().get(i));
-                }
-            }
-            originalFC.setFeatures(visibleFeatures);
-        }
-        ArrayList<String> labels = new ArrayList<>();
-        for (Dimension d : Dimension.getDimensions()) {
-            labels.add(d.labelString);
-        }
-        originalFC.setDimensionLabels(labels);
-        return originalFC;
-    }
-
-    private void setupMissionTimer(Scene scene) {
-        double width = 900;
-        double height = 200;
-        double iconFitWidth = 30;
-        if (is4k) {
-            width = 1800;
-            height = 400;
-            iconFitWidth = 60;
-        }
-        missionTimerX = MissionTimerXBuilder.create()
-            .title("Trinity Mission Time")
-            .startTime(0)
-            .timeFrame(java.time.Duration.ofMinutes(14).getSeconds())
-            .maxSize(width, height)
-            .prefSize(width, height)
-            .iconFitWidth(iconFitWidth)
-            .padding(Insets.EMPTY)
-            .ringColor(Color.STEELBLUE.deriveColor(1, 1, 1, 0.333))
-            .backgroundColor(Color.BLACK.deriveColor(1, 1, 1, 0.333))
-            .itemInnerColor(Color.CYAN.deriveColor(1, 1, 1, 0.75))
-            .clockColor(Color.ALICEBLUE)
-            .ringBackgroundColor(Color.BLACK.deriveColor(1, 1, 1, 0.333))
-            .titleColor(Color.ALICEBLUE)
-            .build();
-        StackPane.setAlignment(missionTimerX, Pos.BOTTOM_CENTER);
-        missionTimerX.setPickOnBounds(false);
-        missionTimerX.updateTime(0, "1");
-        scene.addEventHandler(TimelineEvent.TIMELINE_SET_VISIBLE, e -> missionTimerX.setVisible((boolean) e.object));
-        scene.addEventHandler(TimelineEvent.TIMELINE_CLEAR_ITEMS, e -> missionTimerX.clearItems());
-        scene.addEventHandler(TimelineEvent.TIMELINE_ADD_ITEMS, e -> missionTimerX.addItems((List<Item>) e.object));
-        scene.addEventHandler(TimelineEvent.TIMELINE_SET_INITIALTIME, e -> missionTimerX.setInitialTime((LocalDateTime) e.object));
-        missionTimerX.setVisible(false);
-        scene.addEventHandler(TimelineEvent.TIMELINE_UPDATE_ANIMATIONTIME, e -> {
-            missionTimerX.updateTime(timelineAnimation.getTimeFromStart(),
-                timelineAnimation.getStringCurrentPropRate());
-        });
-        // EVENT HANDLER TO CHANGE THE PROPAGATION RATE ON SCREEN
-        scene.addEventHandler(TimelineEvent.TIMELINE_GET_CURRENT_PROP_RATE, e -> {
-            missionTimerX.updatePropRate(timelineAnimation.getStringCurrentPropRate());
-        });
-        //let timelineAnimation control missionTimerX
-        missionTimerX.internalClock = false; //Disable internal ticking...
-        missionTimerX.start();
-        timelineAnimation = new TimelineAnimation();
-        timelineAnimation.setScene(scene);
-        timelineAnimation.start();
-
+    private void intro() {
+//        Platform.runLater(() -> {
+            intro = new Timeline(
+                new KeyFrame(Duration.seconds(0.0), new KeyValue(mainNavMenu.opacityProperty(), 0.0)),
+                new KeyFrame(Duration.seconds(0.1), e -> {
+                    circleSpinner.spin(false);
+                    circleSpinner.fadeBusy(true);
+                    animatedConsoleText.setOpacity(1.0);
+                    animatedConsoleText.setText("> ");
+                    animatedConsoleText.setVisible(true);
+                    centerStack.getChildren().add(centerStack.getChildren().size() - 1, mainNavMenu);
+                }),
+                new KeyFrame(Duration.seconds(0.5), e -> animatedConsoleText.animate(">Trinity")),
+                new KeyFrame(Duration.seconds(2.0), e -> animatedConsoleText.animate(">Hyperdimensional Visualization")),
+                new KeyFrame(Duration.seconds(3.5), new KeyValue(animatedConsoleText.opacityProperty(), 1.0)),
+                new KeyFrame(Duration.seconds(5.0), new KeyValue(animatedConsoleText.opacityProperty(), 0.0)),
+                new KeyFrame(Duration.seconds(5.1), e -> animatedConsoleText.setVisible(false)),
+                new KeyFrame(Duration.seconds(5.1), e -> animatedConsoleText.animate(" ")),
+                new KeyFrame(Duration.seconds(5.1), new KeyValue(mainNavMenu.opacityProperty(), 1.0))
+            );
+            intro.play();
+//        });
     }
 
     private void fadeOutConsole(long timeMS) {
@@ -1066,7 +349,6 @@ public class App extends Application {
         fade.setOnFinished(f -> animatedConsoleText.setText("> "));
         fade.play();
     }
-
     private void fadeInConsole(long timeMS) {
         animatedConsoleText.setAnimationTimeMS(30);  //default is 30ms
         FadeTransition fade = new FadeTransition(Duration.millis(timeMS), animatedConsoleText);
@@ -1075,24 +357,6 @@ public class App extends Application {
         fade.setOnFinished(f -> animatedConsoleText.setText("> "));
         fade.play();
     }
-
-    private void intro() {
-        Platform.runLater(() -> {
-            animatedConsoleText.setOpacity(1.0);
-            animatedConsoleText.setText("> ");
-            animatedConsoleText.setVisible(true);
-            intro = new Timeline(
-                new KeyFrame(Duration.seconds(0.1), e -> animatedConsoleText.animate(">Trinity")),
-                new KeyFrame(Duration.seconds(2.0), e -> animatedConsoleText.animate(">Hyperdimensional Visualization")),
-                new KeyFrame(Duration.seconds(3.5), new KeyValue(animatedConsoleText.opacityProperty(), 1.0)),
-                new KeyFrame(Duration.seconds(5.0), new KeyValue(animatedConsoleText.opacityProperty(), 0.0)),
-                new KeyFrame(Duration.seconds(5.1), e -> animatedConsoleText.setVisible(false)),
-                new KeyFrame(Duration.seconds(5.1), e -> animatedConsoleText.animate(" "))
-            );
-            intro.play();
-        });
-    }
-
     /**
      * Key handler for various keyboard shortcuts
      */
@@ -1125,12 +389,8 @@ public class App extends Application {
                 new ApplicationEvent(ApplicationEvent.SHOW_PROJECTOR_PANE));
         }
         if (e.isAltDown() && e.isControlDown() && e.getCode().equals(KeyCode.I)) {
-            if (!pathPane.getChildren().contains(imageInspectorPane)) {
-                pathPane.getChildren().add(imageInspectorPane);
-                imageInspectorPane.slideInPane();
-            } else {
-                imageInspectorPane.show();
-            }
+            stage.getScene().getRoot().fireEvent(
+                new ApplicationEvent(ApplicationEvent.SHOW_IMAGE_INSPECTION));
         }
         if (e.isAltDown() && e.isControlDown() && e.getCode().equals(KeyCode.H)) {
             stage.getScene().getRoot().fireEvent(
@@ -1142,7 +402,12 @@ public class App extends Application {
         }
         if (e.isAltDown() && e.getCode().equals(KeyCode.N)) {
             matrixShowing = !matrixShowing;
-            if (!matrixShowing) {
+            if (matrixShowing) {
+                MatrixOverlay.on.set(true);
+                matrixOverlay.matrixOn.run();
+            } else {
+                MatrixOverlay.on.set(false);
+                matrixOverlay.matrixOff.run();                
                 intro();
             }
         }
@@ -1155,12 +420,16 @@ public class App extends Application {
     }
 
     private void shutdown(boolean now) {
-        if (null != missionTimerX) {
-            missionTimerX.stop();
-        }
-        if (null != feed) {
-            feed.disconnect(true); //if no connection does nothing
-        }
+        animatedConsoleText.getScene().getRoot().fireEvent(
+            new ZeroMQEvent(ZeroMQEvent.ZEROMQ_TERMINATE_CONNECTION, null));        
+//        if (null != feed) {
+//            feed.disconnect(true); //if no connection does nothing
+//        }
+        animatedConsoleText.getScene().getRoot().fireEvent(
+            new MissionTimerXEvent(MissionTimerXEvent.MISSION_TIMER_SHUTDOWN)); 
+//        if (null != missionTimerX) {
+//            missionTimerX.stop();
+//        }
         if (now) {
             System.exit(0);
         } else {
@@ -1181,7 +450,7 @@ public class App extends Application {
     private void parseCommandLine() {
         Parameters parameters = getParameters();
         namedParameters = parameters.getNamed();
-        unnamedParameters = parameters.getUnnamed();
+//        unnamedParameters = parameters.getUnnamed();
 
         if (!namedParameters.isEmpty()) {
             LOG.info("NamedParameters :");
@@ -1189,10 +458,10 @@ public class App extends Application {
                 LOG.info("\t{} : {}", entry.getKey(), entry.getValue());
             });
         }
-        if (!unnamedParameters.isEmpty()) {
-            LOG.info("UnnamedParameters :");
-            unnamedParameters.forEach(System.out::println);
-        }
+//        if (!unnamedParameters.isEmpty()) {
+//            LOG.info("UnnamedParameters :");
+//            unnamedParameters.forEach(System.out::println);
+//        }
     }
 
     //<editor-fold defaultstate="collapsed" desc="Static Accessors">
